First I added a lot of new directives so that the code had the right directives to run efficently.
Then I went through the internet to find a new way to add a new item into a vector because the one that was being used didn't work so I used the .pushback to enter Despicable Me 4 to the end of the vector.

Then the arrows were the wrong way for the endl on the announcement so I switched those around.

I then switch the for loop to make it so it went off the size of the movies list and couldn't go over. Then had it send out each movie in the list of movies in the vector using the for loop i.

(Random stuff stayed the same because nothing was wrong)

Then connected the pick a number command into a new variable called winner so that the movie being picked at random could be in one single variable.

Then we needed to take the whole word of the one that it chose and turn it all into uppercase letters so I used the beginning middle and the end to do so.

Then it takes the winner and annonces the new all caps of the randomly chose movie and has to say that it is the winner.